requirements.txt
smart-open
requests
python-dotenv